Visit ReturnStmt in TBR
