+++ include/clad/Differentiator/ReverseModeVisitor.h
@@ -215,10 +215,6 @@ namespace clad {
     /// to avoid recomputiation.
     bool UsefulToStoreGlobal(clang::Expr* E);
 
-    /// For an expr E, decides if we should recompute it or store it.
-    /// This is the central point for checkpointing.
-    bool ShouldRecompute(const clang::Expr* E);
-
     /// Builds a variable declaration and stores it in the function
     /// global scope.
     ///
