+++ lib/Differentiator/CladUtils.cpp
@@ -1488,5 +1488,27 @@ namespace clad {
       } while (E->IgnoreImplicit() != E);
       return false;
     }
+
+    /// Called in ShouldRecompute. In CUDA, to access a current thread/block id
+    /// we use functions that do not change the state of any variable, since no
+    /// point to store the value.
+    static bool isCUDABuiltInIndex(const Expr* E) {
+      const clang::Expr* B = E->IgnoreImplicit();
+      if (const auto* pseudoE = llvm::dyn_cast<PseudoObjectExpr>(B)) {
+        if (const auto* opaqueE =
+                llvm::dyn_cast<OpaqueValueExpr>(pseudoE->getSemanticExpr(0))) {
+          const Expr* innerE = opaqueE->getSourceExpr()->IgnoreImplicit();
+          QualType innerT = innerE->getType();
+          if (innerT.isConstQualified())
+            return true;
+        }
+      }
+      return false;
+    }
+
+    bool ShouldRecompute(const Expr* E, const ASTContext& C) {
+      return !(utils::ContainsFunctionCalls(E) || E->HasSideEffects(C)) ||
+             isCUDABuiltInIndex(E);
+    }
   } // namespace utils
 } // namespace clad
